Refactor reusable code to package object
